+++ priv/repo/migrations/20260902110000_prefix_primary_key_names.exs
@@ -1,0 +1,5 @@
+defmodule Boruta.Repo.Migrations.PrefixPrimaryKeyNames do
+  use Ecto.Migration
+
+  use Boruta.Migrations.PrefixPrimaryKeyNames
+end