+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) according to OAuth / OpenID connect specifications, changes may break in order to comply with those.
 
+## [Unreleased]
+
+- Note that you must run the `boruta.gen.migration` task to rename the primary key constraints.
+
+### Fixed
+
+- primary keys and the `oauth_clients_scopes` sequence carry the `oauth_` prefixed names; duplicate client and scope ids are reported as changeset errors whichever name the constraint has
+
 ## [2.3.8] - 2026-08-01
 
 ### Fixed